fix: ensure that test_ygg_metadata_foreach works in any case
YggMetadata uses a GHashTable to keep the various key/value pairs, so
there is no defined order for them. This means that when using
ygg_metadata_foreach() the key/value pairs will be provided in an
undefined order.

The current test seems to rely on the order of the keys, i.e. in the
same order passed to create the GVariant. To make the test work
regardless of the internals of glib, collect the list of key/value pairs
into a list, and check for all the pairs explicitly.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>